tests 'ci skip'

+++ tests/ProcessTest.php
@@ -52,7 +52,9 @@ class ProcessTest extends PHPUnit_Framework_TestCase
         });
 
         $cache = new \Jenner\SimpleFork\Cache\SharedMemory();
-        $cache->delete('test');
+        if($cache->has('test')){
+            $cache->delete('test');
+        }
         $process->start();
         sleep(1);
         $this->assertEquals('test', $cache->get('test'));
